Rudimentary hwcap support
